Adjust more tests.
The only ones failing now are for the accept_only method, which doesn't
make much sense.  The only place it's used in this code base is to
support the broken idea that the Content-Type header should take
precedence over the Accept header in determining what serialization
format should be used for the response.