Responses vary based on Accept and Content-Type
Previously the response was only stated to vary based on the
Content-Type header, unless that wasn't present then it would only list
the Accept header.  This is wrong for two reasons.

First, the Accept header should take precedence for the response, so the
presence of a Content-Type header doesn't render the Accept header
meaningless.

Second, even if the Content-Type header should take precedence, it would
need to be listed in the Vary header even if not present.  Otherwise a
cached response to a request with no Content-Type header could be
delivered for a later response that did include a Content-Type header.